Added the concept of a SceneManager
A scene can contain a lot of scene managers. It will come with the DebugManager and NotificationManager preinstalled, but can incorporate custom ones as well. These managers have all the capabilities of a scene, but are layered on top of it

+++ src/main/java/com/ejo/ui/element/base/IInteractable.java
@@ -8,6 +8,6 @@ public interface IInteractable {
 
     void onMouseClick(int button, int action, int mods, Vector mousePos);
 
-    void onMouseScroll(int scroll, Vector mousePos);
+    void onMouseScroll(double scroll, Vector mousePos);
 
 }
